Upgrade Engine to C++20

+++ Engine/runtime/render_a/src/sys/d3d/d3d_renderblock.cpp
@@ -349,12 +349,12 @@ ILTStream &operator>>(ILTStream &cStream, SRBOccluder &cPoly)
 ILTStream &operator>>(ILTStream &cStream, SRBLightGroup &cLightGroup)
 {
 	uint16 nLength;
-	cStream >> (uint16)nLength;
+	cStream >> nLength;
 	cLightGroup.m_nID = 0;
 	for (; nLength; --nLength)
 	{
 		uint8 nNextChar;
-		cStream >> (uint8)nNextChar;
+		cStream >> nNextChar;
 		cLightGroup.m_nID *= 31;
 		cLightGroup.m_nID += (uint32)nNextChar;
 	}
@@ -362,7 +362,7 @@ ILTStream &operator>>(ILTStream &cStream, SRBLightGroup &cLightGroup)
 	cStream >> cLightGroup.m_vColor;
 
 	uint32 nDataLength;
-	cStream >> (uint32)nDataLength;
+	cStream >> nDataLength;
 	LT_MEM_TRACK_ALLOC(cLightGroup.m_aVertexIntensities.resize(nDataLength, 0), LT_MEM_TYPE_RENDER_LIGHTGROUP);
 	cStream.Read(static_cast<void*>(&(*cLightGroup.m_aVertexIntensities.begin())), nDataLength);
 
@@ -526,9 +526,9 @@ bool CD3D_RenderBlock::Load(ILTStream *pStream)
 			}
 
 			// Load the lightmap
-			*pStream >> (uint32)cCurSection.m_nLightmapWidth;
-			*pStream >> (uint32)cCurSection.m_nLightmapHeight;
-			*pStream >> (uint32)cCurSection.m_nLightmapSize;
+			*pStream >> cCurSection.m_nLightmapWidth;
+			*pStream >> cCurSection.m_nLightmapHeight;
+			*pStream >> cCurSection.m_nLightmapSize;
 			if (cCurSection.m_nLightmapSize)
 			{
 				LT_MEM_TRACK_ALLOC(cCurSection.m_pLightmapData = new uint8[cCurSection.m_nLightmapSize],LT_MEM_TYPE_RENDER_LIGHTMAP);
@@ -544,7 +544,7 @@ bool CD3D_RenderBlock::Load(ILTStream *pStream)
 	}
 
 	// Get the vertex data
-	*pStream >> (uint32)m_nVertexCount;
+	*pStream >> m_nVertexCount;
 	if (m_nVertexCount)
 	{
 		LT_MEM_TRACK_ALLOC(m_aVertices = new SRBVertex[m_nVertexCount],LT_MEM_TYPE_RENDER_WORLD);
@@ -563,7 +563,7 @@ bool CD3D_RenderBlock::Load(ILTStream *pStream)
 	}
 
 	// Get the triangle count
-	*pStream >> (uint32)m_nTriCount;
+	*pStream >> m_nTriCount;
 	if (m_nTriCount)
 	{
 		LT_MEM_TRACK_ALLOC(m_aIndices = new uint16[m_nTriCount * 3],LT_MEM_TYPE_RENDER_WORLD);
